Add shorthand import of all algorithms at package init
In sake of the experience using this tool, an access factory function for obtaining all the implemented algorithms was added in to the initor of the `algorithms` package. Now it's possible to just import the function and provide the State Space instance (representing the problem which solution is desired) to it.

Also added the integration of this in to the `main.py` module.