Add Greedy Search algorithm
This algorithm provides a generalization of Gradient Search. It is based on ordered searching of State Space minding a heuristic function. In this case, the heuristic function is meant to be an euclidean distance between currently searched node of the graph (with uniform length of the edge) and the goal node (final field; state).

This algorithm does not get stuck in a local extreme, but its complexity (in terms of memory and time) reaches much higher levels than Gradient Search.